Refactor code and rename package
In this commit, the package and related files have been renamed from 'oscharkoFormValidation' to 'oscharkoValidate' to make it more concise. Also, the string validation rule implementation was refactored in the 'stringIsTheSameAs.ts' file and unnecessary imports were removed. Lastly, versions of dependencies in the 'package.json' file have been updated and set to exact versions to improve build stability.